requirements.txt
annotated-types==0.6.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
Babel==2.14.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.0
cryptography==41.0.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0
django-allauth==0.59.0
django-appconf==1.0.6
django-compressor==4.4
django-extensions==3.2.3
et-xmlfile==1.1.0
executing==2.0.1
idna==3.6
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
JPype1==1.5.0
jupyter_client==8.6.0
jupyter_core==5.5.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
pytailwindcss==0.2.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
pyzmq==25.1.2
rcssmin==1.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tabula-py==2.9.0
tornado==6.4
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.12
